VB-119: move fetch products to Filter component
