+++ packages/frontend/pages/items/filter.tsx
@@ -1,16 +1,6 @@
 import { Filter, Layout } from '@components';
-import { useRouter } from 'next/router';
 import { wrapper } from 'store';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
-import { useAppDispatch, useTypedSelector } from '@hooks';
-import { fetchProducts } from 'store/product';
-import { useEffect } from 'react';
-import type { ProductType } from '@vse-bude/shared';
-
-interface RequestOptions {
-  limit?: number;
-  type?: ProductType;
-}
 
 export const getServerSideProps = wrapper.getServerSideProps(
   (_store) =>
@@ -21,22 +11,10 @@ export const getServerSideProps = wrapper.getServerSideProps(
     }),
 );
 
-const FilteredPage = () => {
-  const router = useRouter();
-  const query = router.query as unknown as string;
-  const { list } = useTypedSelector((store) => store.product);
-  const filter: RequestOptions = Object.fromEntries(new URLSearchParams(query));
-  const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    dispatch(fetchProducts({ limit: filter.limit, type: filter.type }));
-  }, [dispatch, filter]);
-
-  return (
+const FilteredPage = () => (
     <Layout title="Filtered posts">
-      <Filter lots={list} />
+      <Filter />
     </Layout>
   );
-};
 
 export default FilteredPage;
